Move isRoleValid and areProjectRolesValid to shared
