+++ packages/enterprise/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  moduleFileExtensions: ["ts", "js", "node"],
+  moduleFileExtensions: ["ts", "js"],
   transform: {
     "^.+\\.(ts|tsx)$": "@swc/jest",
   },
