+++ packages/shared/jest.config.js
@@ -1,10 +1,7 @@
 module.exports = {
-  moduleFileExtensions: ["ts", "js", "node"],
+  moduleFileExtensions: ["ts", "js"],
   transform: {
     "^.+\\.(ts|tsx)$": "@swc/jest",
   },
   testMatch: ["**/test/**/*.test.(ts|js)"],
-  moduleNameMapper: {
-    "^axios$": "axios/dist/axios.js",
-  },
 };
